Move ReadExactExt to a util module
