+++ src/lib.rs
@@ -8,3 +8,4 @@ extern crate uuid;
 
 pub mod packet;
 pub mod types;
+mod util;
